separate functions into own file
