+++ src/every.js
@@ -1,0 +1,10 @@
+const reduce = require('./reduce')
+
+module.exports = function(collection, iterator, callback) {
+  return reduce(collection, (trueSoFar, value) => (trueSoFar && iterator(value)), true)
+    .then((result) => (callback ? callback(!!result) : !!result))
+    .catch((error) => {
+      if (callback) return callback(error)
+      throw error
+    })
+}
